picture_frame/requirements.txt
PySimpleGUI-4.60.4-py3-none-any.whl
pillow
pillow-heif
